Updated documentation and docker image
Updated README instructions to involve the docker build and run steps as this will be the preferred way to run Freshli-Cli.
Updated the Dockerfile to pull from the dotnet-focal image as the regular dotnet image started having cert issues today. I will look more into this later.
As is, this will not run local until the LibGit2Sharp package is updated in Freshli. Otherwise, Freshli will fail running on the linux based dotnet image.